V1

+++ app/src/main/java/com/snakeindustry/snakemultiplayer/Snake/model/SnakeCell.java
@@ -5,26 +5,26 @@ package com.snakeindustry.snakemultiplayer.Snake.model;
  */
 public class SnakeCell {
 
-    private int x,y;
+    private double x,y;
 
-    public SnakeCell(int x,int y){
-        this.x=x;
-        this.y=y;
+    public SnakeCell(double x, double y) {
+        this.x = x;
+        this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 }
